+++ e310x-hal/src/asynch.rs
@@ -1,0 +1,9 @@
+//! Asynchronous HAL for the E310x family of microcontrollers
+//!
+//! This is an implementation of the [`embedded-hal-async`] traits for the E310x
+//! family of microcontrollers.
+
+#![deny(missing_docs)]
+
+pub mod digital;
+pub mod prelude;