+++ e310x-hal/src/asynch/prelude.rs
@@ -1,0 +1,2 @@
+//! Prelude
+pub use embedded_hal_async::digital::Wait as _eha_Wait;