+++ e310x-hal/src/gpio.rs
@@ -558,6 +558,7 @@ macro_rules! gpio {
                     }
                 }
 
+                #[cfg(not(feature = "async"))]
                 impl<MODE> ErrorType for $PXi<Input<MODE>> {
                     type Error = Infallible;
                 }